refactor(lms): modify comments - (vo domain 구분짓는 이유가 없음)https://github.com/next-step/java-lms/pull/580#discussion_r1818280246
