+++ src/main/java/nextstep/courses/domain/session/CoverImage.java
@@ -1,8 +1,8 @@
-package nextstep.courses.domain.vo.session;
+package nextstep.courses.domain.session;
 
-import nextstep.courses.domain.vo.session.image.ImageFileSize;
-import nextstep.courses.domain.vo.session.image.ImageType;
-import nextstep.courses.domain.vo.session.image.ImageSize;
+import nextstep.courses.domain.session.image.ImageFileSize;
+import nextstep.courses.domain.session.image.ImageType;
+import nextstep.courses.domain.session.image.ImageSize;
 
 import java.util.Objects;
