+++ src/main/java/nextstep/courses/domain/session/DateRange.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.vo.session;
+package nextstep.courses.domain.session;
 
 import nextstep.courses.EndBeforeStartException;
