+++ src/main/java/nextstep/courses/domain/session/image/ImageSize.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.vo.session.image;
+package nextstep.courses.domain.session.image;
 
 import nextstep.courses.ImageWidthHeightRatioMismatchException;
 import nextstep.courses.WidthHeightMinimumException;
