+++ src/test/java/nextstep/courses/domain/session/StudentsTest.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.vo.session;
+package nextstep.courses.domain.session;
 
 import nextstep.users.domain.NsUser;
 import nextstep.users.domain.NsUserTest;
